broswer_email_cs is TypeScript
